// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** The type of a cursor, an opaque string used for walking connections */
  Cursor: { input: any; output: any };
  /** An RFC-3339 compliant Full Date Scalar */
  Date: { input: any; output: any };
  /** A slightly refined version of RFC-3339 compliant DateTime Scalar */
  DateTime: { input: any; output: any };
  /** A 64-bit signed integer */
  Long: { input: any; output: any };
  /** An Float scalar that must be greater than or equal to zero */
  NonNegativeFloat: { input: any; output: any };
  /** An Int scalar that must be greater than or equal to zero */
  NonNegativeInt: { input: any; output: any };
  /** An Int scalar that must be a positive value */
  PositiveInt: { input: any; output: any };
};
// ...
export type Quantity = {
  __typename?: "Quantity";
  quantity: Scalars["Float"]["output"];
  units?: Maybe<UnitOfMeasure>;
};
// ...
export enum RecognizedRangeType {
  Item = "ITEM",
  NewItem = "NEW_ITEM",
  NewUnit = "NEW_UNIT",
  Quantity = "QUANTITY",
  Unit = "UNIT",
  Unknown = "UNKNOWN",
}
// ...
export type UnitOfMeasure = Node & {
  __typename?: "UnitOfMeasure";
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
};